Resolves #2 - Added classes subdirectory
