+++ classes/index.js
@@ -1,0 +1,7 @@
+const AntiraidSettings = require('./antiraid-settings');
+const GuildAntiraidSettingsSchema = require('./guild-antiraid-settings-schema');
+
+module.exports = {
+	AntiraidSettings,
+	GuildAntiraidSettingsSchema
+};
